requirements.txt
alembic==1.18.4
click==8.3.1
fastapi==0.133.0
httptools==0.7.1
pyyaml==6.0.3
starlette==0.52.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
pytest==9.0.2
pytest-asyncio==1.3.0
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-telegram-bot==22.6
sniffio==1.3.1
SQLAlchemy==2.0.46
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
